Replaced Dizzy (for now) with Dummy
Realized that I should probably be using a non complicated model and animations for prototyping the character controller. So. Dizzy has been replaced with Dummy. Animated in-engine instead of Blender to keep the workflow simple.

Class Character now properly checks if an AnimationPlayer's animation_library is empty/null.

Added new node, CharacterAnimationTreeExpressionBase, responsible for interacting with a Character's AnimationTree node.

Both aforementioned and AnimationTree node added to the default Character scene and is subsequently a part of inherited scenes (dizzy_character.tscn and dummy_character.tscn).

For now, the proposed default AnimationTree StateMachine is contained in dummy_character.tscn. May be moved to the default character.tscn scene to make character creation easier.

Moved HitboxAnimationPlayer out of inherited scene and directly into the default scene character_model.tscn.

Added debug toggle debug_no_animations to allow the prevention of animations for a PlayableCharacter.

Added "Landing" as an animation state on playable_character.tscn.

All AnimationStates on playable_character.tscn are now updated to interact with an AnimationTree on the given Character instead of directly with an AnimationPlayer.

Dizzy model has been completely remade, but she is not finished otherwise.