Unfinished Character Combo Editor Addon + Entities and Their AI
...Lot's of things going on in this commit because I forgot to commit the Character Combo Editor addon before starting something else. It's currently unfinished and as such will be entirely undocumented because it is highly likely to change in the future.

Instead of that, I decided to work on a simple Entity (the name of the Enemies) that can easily be modified via an inherited scene and modifications to the base behaviour tree.

Added ExploreEditorTheme addon for more seamless visuals of any addons to be made in the future.

Added the Entity node, responsible for managing all components of a given Entity. The Entity scene is the base scene for all entities (Kilobyte, Megabyte, Gigabyte or Terabyte) to inherit from to make modification of Entity behaviour much easier and modular. The general functionality is simple. The BehaviourTree acts like the brain of the Entity. All behaviours that will control the Entity's movement are found in the BehaviourTree. the BehaviourTree's Blackboard will then be responsible for relaying values such as input_direction to the StateMachine, which will make the Entity behave accordingly.